Type Definition stm32f0::stm32f0x1::adc::isr::ADRDY_R

type ADRDY_R = R<bool, ADRDY_A>;

Reader of field ADRDY

Implementations

impl ADRDY_R

pub fn variant(&self) -> ADRDY_A

Get enumerated values variant

pub fn is_not_ready(&self) -> bool

Checks if the value of the field is NOTREADY

pub fn is_ready(&self) -> bool

Checks if the value of the field is READY