Type Definition stm32f0::stm32f0x1::adc::smpr::SMP_R

type SMP_R = R<u8, SMP_A>;

Reader of field SMP

Implementations

impl SMP_R

pub fn variant(&self) -> SMP_A

Get enumerated values variant

pub fn is_cycles1_5(&self) -> bool

Checks if the value of the field is CYCLES1_5

pub fn is_cycles7_5(&self) -> bool

Checks if the value of the field is CYCLES7_5

pub fn is_cycles13_5(&self) -> bool

Checks if the value of the field is CYCLES13_5

pub fn is_cycles28_5(&self) -> bool

Checks if the value of the field is CYCLES28_5

pub fn is_cycles41_5(&self) -> bool

Checks if the value of the field is CYCLES41_5

pub fn is_cycles55_5(&self) -> bool

Checks if the value of the field is CYCLES55_5

pub fn is_cycles71_5(&self) -> bool

Checks if the value of the field is CYCLES71_5

pub fn is_cycles239_5(&self) -> bool

Checks if the value of the field is CYCLES239_5