Type Definition stm32f0::stm32f0x1::comp::csr::COMP2OUTSEL_R

type COMP2OUTSEL_R = R<u8, COMP2OUTSEL_A>;

Reader of field COMP2OUTSEL

Implementations

impl COMP2OUTSEL_R

pub fn variant(&self) -> COMP2OUTSEL_A

Get enumerated values variant

pub fn is_no_selection(&self) -> bool

Checks if the value of the field is NOSELECTION

pub fn is_timer1break_input(&self) -> bool

Checks if the value of the field is TIMER1BREAKINPUT

pub fn is_timer1input_capture1(&self) -> bool

Checks if the value of the field is TIMER1INPUTCAPTURE1

pub fn is_timer1ocref_clear_input(&self) -> bool

Checks if the value of the field is TIMER1OCREFCLEARINPUT

pub fn is_timer2input_capture4(&self) -> bool

Checks if the value of the field is TIMER2INPUTCAPTURE4

pub fn is_timer2ocref_clear_input(&self) -> bool

Checks if the value of the field is TIMER2OCREFCLEARINPUT

pub fn is_timer3input_capture1(&self) -> bool

Checks if the value of the field is TIMER3INPUTCAPTURE1

pub fn is_timer3ocref_clear_input(&self) -> bool

Checks if the value of the field is TIMER3OCREFCLEARINPUT