Type Definition stm32f0::stm32f0x1::i2c1::cr1::DNF_R

type DNF_R = R<u8, DNF_A>;

Reader of field DNF

Implementations

impl DNF_R

pub fn variant(&self) -> DNF_A

Get enumerated values variant

pub fn is_no_filter(&self) -> bool

Checks if the value of the field is NOFILTER

pub fn is_filter1(&self) -> bool

Checks if the value of the field is FILTER1

pub fn is_filter2(&self) -> bool

Checks if the value of the field is FILTER2

pub fn is_filter3(&self) -> bool

Checks if the value of the field is FILTER3

pub fn is_filter4(&self) -> bool

Checks if the value of the field is FILTER4

pub fn is_filter5(&self) -> bool

Checks if the value of the field is FILTER5

pub fn is_filter6(&self) -> bool

Checks if the value of the field is FILTER6

pub fn is_filter7(&self) -> bool

Checks if the value of the field is FILTER7

pub fn is_filter8(&self) -> bool

Checks if the value of the field is FILTER8

pub fn is_filter9(&self) -> bool

Checks if the value of the field is FILTER9

pub fn is_filter10(&self) -> bool

Checks if the value of the field is FILTER10

pub fn is_filter11(&self) -> bool

Checks if the value of the field is FILTER11

pub fn is_filter12(&self) -> bool

Checks if the value of the field is FILTER12

pub fn is_filter13(&self) -> bool

Checks if the value of the field is FILTER13

pub fn is_filter14(&self) -> bool

Checks if the value of the field is FILTER14

pub fn is_filter15(&self) -> bool

Checks if the value of the field is FILTER15