Type Definition stm32f0::stm32f0x1::rcc::cfgr::PLLMUL_R

type PLLMUL_R = R<u8, PLLMUL_A>;

Reader of field PLLMUL

Implementations

impl PLLMUL_R

pub fn variant(&self) -> PLLMUL_A

Get enumerated values variant

pub fn is_mul2(&self) -> bool

Checks if the value of the field is MUL2

pub fn is_mul3(&self) -> bool

Checks if the value of the field is MUL3

pub fn is_mul4(&self) -> bool

Checks if the value of the field is MUL4

pub fn is_mul5(&self) -> bool

Checks if the value of the field is MUL5

pub fn is_mul6(&self) -> bool

Checks if the value of the field is MUL6

pub fn is_mul7(&self) -> bool

Checks if the value of the field is MUL7

pub fn is_mul8(&self) -> bool

Checks if the value of the field is MUL8

pub fn is_mul9(&self) -> bool

Checks if the value of the field is MUL9

pub fn is_mul10(&self) -> bool

Checks if the value of the field is MUL10

pub fn is_mul11(&self) -> bool

Checks if the value of the field is MUL11

pub fn is_mul12(&self) -> bool

Checks if the value of the field is MUL12

pub fn is_mul13(&self) -> bool

Checks if the value of the field is MUL13

pub fn is_mul14(&self) -> bool

Checks if the value of the field is MUL14

pub fn is_mul15(&self) -> bool

Checks if the value of the field is MUL15

pub fn is_mul16(&self) -> bool

Checks if the value of the field is MUL16

pub fn is_mul16x(&self) -> bool

Checks if the value of the field is MUL16X