Type Definition stm32f0::stm32f0x1::rcc::cfgr2::PREDIV_R

type PREDIV_R = R<u8, PREDIV_A>;

Reader of field PREDIV

Implementations

impl PREDIV_R

pub fn variant(&self) -> PREDIV_A

Get enumerated values variant

pub fn is_div1(&self) -> bool

Checks if the value of the field is DIV1

pub fn is_div2(&self) -> bool

Checks if the value of the field is DIV2

pub fn is_div3(&self) -> bool

Checks if the value of the field is DIV3

pub fn is_div4(&self) -> bool

Checks if the value of the field is DIV4

pub fn is_div5(&self) -> bool

Checks if the value of the field is DIV5

pub fn is_div6(&self) -> bool

Checks if the value of the field is DIV6

pub fn is_div7(&self) -> bool

Checks if the value of the field is DIV7

pub fn is_div8(&self) -> bool

Checks if the value of the field is DIV8

pub fn is_div9(&self) -> bool

Checks if the value of the field is DIV9

pub fn is_div10(&self) -> bool

Checks if the value of the field is DIV10

pub fn is_div11(&self) -> bool

Checks if the value of the field is DIV11

pub fn is_div12(&self) -> bool

Checks if the value of the field is DIV12

pub fn is_div13(&self) -> bool

Checks if the value of the field is DIV13

pub fn is_div14(&self) -> bool

Checks if the value of the field is DIV14

pub fn is_div15(&self) -> bool

Checks if the value of the field is DIV15

pub fn is_div16(&self) -> bool

Checks if the value of the field is DIV16