Type Definition stm32f0::stm32f0x1::spi1::cr2::DS_R

type DS_R = R<u8, DS_A>;

Reader of field DS

Implementations

impl DS_R

pub fn variant(&self) -> Variant<u8, DS_A>

Get enumerated values variant

pub fn is_four_bit(&self) -> bool

Checks if the value of the field is FOURBIT

pub fn is_five_bit(&self) -> bool

Checks if the value of the field is FIVEBIT

pub fn is_six_bit(&self) -> bool

Checks if the value of the field is SIXBIT

pub fn is_seven_bit(&self) -> bool

Checks if the value of the field is SEVENBIT

pub fn is_eight_bit(&self) -> bool

Checks if the value of the field is EIGHTBIT

pub fn is_nine_bit(&self) -> bool

Checks if the value of the field is NINEBIT

pub fn is_ten_bit(&self) -> bool

Checks if the value of the field is TENBIT

pub fn is_eleven_bit(&self) -> bool

Checks if the value of the field is ELEVENBIT

pub fn is_twelve_bit(&self) -> bool

Checks if the value of the field is TWELVEBIT

pub fn is_thirteen_bit(&self) -> bool

Checks if the value of the field is THIRTEENBIT

pub fn is_fourteen_bit(&self) -> bool

Checks if the value of the field is FOURTEENBIT

pub fn is_fifteen_bit(&self) -> bool

Checks if the value of the field is FIFTEENBIT

pub fn is_sixteen_bit(&self) -> bool

Checks if the value of the field is SIXTEENBIT