Type Definition stm32f0::stm32f0x1::syscfg::exticr2::EXTI7_R

type EXTI7_R = R<u8, EXTI7_A>;

Reader of field EXTI7

Implementations

impl EXTI7_R

pub fn variant(&self) -> Variant<u8, EXTI7_A>

Get enumerated values variant

pub fn is_pe7(&self) -> bool

Checks if the value of the field is PE7

pub fn is_pa7(&self) -> bool

Checks if the value of the field is PA7

pub fn is_pb7(&self) -> bool

Checks if the value of the field is PB7

pub fn is_pc7(&self) -> bool

Checks if the value of the field is PC7

pub fn is_pd7(&self) -> bool

Checks if the value of the field is PD7

pub fn is_pf7(&self) -> bool

Checks if the value of the field is PF7