Type Definition stm32f0::stm32f0x1::syscfg::exticr3::EXTI10_R

type EXTI10_R = R<u8, EXTI10_A>;

Reader of field EXTI10

Implementations

impl EXTI10_R

pub fn variant(&self) -> Variant<u8, EXTI10_A>

Get enumerated values variant

pub fn is_pe10(&self) -> bool

Checks if the value of the field is PE10

pub fn is_pa10(&self) -> bool

Checks if the value of the field is PA10

pub fn is_pb10(&self) -> bool

Checks if the value of the field is PB10

pub fn is_pc10(&self) -> bool

Checks if the value of the field is PC10

pub fn is_pd10(&self) -> bool

Checks if the value of the field is PD10

pub fn is_pf10(&self) -> bool

Checks if the value of the field is PF10