Type Definition stm32f0::stm32f0x1::syscfg::exticr3::EXTI8_R

type EXTI8_R = R<u8, EXTI8_A>;

Reader of field EXTI8

Implementations

impl EXTI8_R

pub fn variant(&self) -> Variant<u8, EXTI8_A>

Get enumerated values variant

pub fn is_pe8(&self) -> bool

Checks if the value of the field is PE8

pub fn is_pa8(&self) -> bool

Checks if the value of the field is PA8

pub fn is_pb8(&self) -> bool

Checks if the value of the field is PB8

pub fn is_pc8(&self) -> bool

Checks if the value of the field is PC8

pub fn is_pd8(&self) -> bool

Checks if the value of the field is PD8

pub fn is_pf8(&self) -> bool

Checks if the value of the field is PF8