Type Definition stm32f0::stm32f0x1::syscfg::exticr4::EXTI13_R

type EXTI13_R = R<u8, EXTI13_A>;

Reader of field EXTI13

Implementations

impl EXTI13_R

pub fn variant(&self) -> Variant<u8, EXTI13_A>

Get enumerated values variant

pub fn is_pe13(&self) -> bool

Checks if the value of the field is PE13

pub fn is_pa13(&self) -> bool

Checks if the value of the field is PA13

pub fn is_pb13(&self) -> bool

Checks if the value of the field is PB13

pub fn is_pc13(&self) -> bool

Checks if the value of the field is PC13

pub fn is_pd13(&self) -> bool

Checks if the value of the field is PD13

pub fn is_pf13(&self) -> bool

Checks if the value of the field is PF13