Type Definition stm32f0::stm32f0x1::syscfg::exticr4::EXTI14_R

type EXTI14_R = R<u8, EXTI14_A>;

Reader of field EXTI14

Implementations

impl EXTI14_R

pub fn variant(&self) -> Variant<u8, EXTI14_A>

Get enumerated values variant

pub fn is_pe14(&self) -> bool

Checks if the value of the field is PE14

pub fn is_pa14(&self) -> bool

Checks if the value of the field is PA14

pub fn is_pb14(&self) -> bool

Checks if the value of the field is PB14

pub fn is_pc14(&self) -> bool

Checks if the value of the field is PC14

pub fn is_pd14(&self) -> bool

Checks if the value of the field is PD14

pub fn is_pf14(&self) -> bool

Checks if the value of the field is PF14