Type Definition stm32f0::stm32f0x1::usart1::cr2::ABRMOD_R

type ABRMOD_R = R<u8, ABRMOD_A>;

Reader of field ABRMOD

Implementations

impl ABRMOD_R

pub fn variant(&self) -> ABRMOD_A

Get enumerated values variant

pub fn is_start(&self) -> bool

Checks if the value of the field is START

pub fn is_edge(&self) -> bool

Checks if the value of the field is EDGE

pub fn is_frame7f(&self) -> bool

Checks if the value of the field is FRAME7F

pub fn is_frame55(&self) -> bool

Checks if the value of the field is FRAME55