Type Definition stm32f0::stm32f0x1::usb::ep4r::EP_TYPE_R

type EP_TYPE_R = R<u8, EP_TYPE_A>;

Reader of field EP_TYPE

Implementations

impl EP_TYPE_R

pub fn variant(&self) -> EP_TYPE_A

Get enumerated values variant

pub fn is_bulk(&self) -> bool

Checks if the value of the field is BULK

pub fn is_control(&self) -> bool

Checks if the value of the field is CONTROL

pub fn is_iso(&self) -> bool

Checks if the value of the field is ISO

pub fn is_interrupt(&self) -> bool

Checks if the value of the field is INTERRUPT